pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tholos"
version = "0.1.0"
description = "Myopic Delirium environment generator and runner"
authors = [{name = "Auto"}]
requires-python = ">=3.11,<3.12"
dependencies = [
    "numpy>=1.26,<2.0",
    "scipy>=1.11,<1.14",
    "numba>=0.59,<0.60",
    "typer==0.9.0",
    "click==8.1.7",
    "pydantic>=2.5,<3.0",
    "jsonschema>=4.21,<5.0",
    "pyyaml>=6.0,<7.0",
    "pandas>=2.1,<3.0",
    "pyarrow>=15,<19",
    "blake3>=0.4,<0.5",
    "matplotlib>=3.7,<4.0",
    "plotly>=5.17,<6.0",
    "imageio>=2.31,<3.0",
    "pillow>=10.0,<11.0"
]

[project.scripts]
tholos = "interfaces.ui_iface.runner.cli:app"

[tool.setuptools.packages.find]
include = ["interfaces*"]
exclude = ["tests*", "runs*", "docs*", "examples*"]
